Implemented New RecentController
