+++ client/src/main/java/application/controller/MainWindowController.java
@@ -1,28 +1,18 @@
 package application.controller;
 
-import application.controller.enums.SubjectRequestResult;
 import application.controller.services.MainConnection;
-import application.controller.services.SubjectRequestService;
 import application.model.Account;
 import application.model.managers.SubjectManager;
 import application.view.ViewFactory;
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.net.URL;
 import java.util.ResourceBundle;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
-import javafx.geometry.Orientation;
 import javafx.scene.control.Button;
 import javafx.scene.control.Label;
-import javafx.scene.control.ScrollBar;
-import javafx.scene.control.ScrollPane;
 import javafx.scene.control.Tab;
 import javafx.scene.control.TabPane;
-import javafx.scene.image.Image;
-import javafx.scene.image.ImageView;
-import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.HBox;
 import javafx.stage.Stage;
@@ -75,70 +65,7 @@ public class MainWindowController extends BaseController implements Initializabl
   private TabPane secondaryTabPane;
 
   @FXML
-  private ImageView tutorAvatarOne;
-
-  @FXML
-  private Label tutorLabelOne;
-
-  @FXML
-  private ImageView tutorAvatarTwo;
-
-  @FXML
-  private Label tutorLabelTwo;
-
-  @FXML
-  private ImageView tutorAvatarThree;
-
-  @FXML
-  private Label tutorLabelThree;
-
-  @FXML
-  private ImageView tutorAvatarFour;
-
-  @FXML
-  private Label tutorLabelFour;
-
-  @FXML
-  private ImageView tutorAvatarFive;
-
-  @FXML
-  private Label tutorLabelFive;
-
-  @FXML
-  private ScrollBar mainRecentScrollBar;
-
-  @FXML
-  private ScrollPane mainRecentScrollPane;
-
-  @FXML
-  private AnchorPane mainRecentScrollContent;
-
-  @FXML
-  private ScrollPane topSubjectsScrollPane;
-
-  @FXML
-  private HBox hboxOne;
-
-  @FXML
-  private HBox hboxTwo;
-
-  @FXML
-  private Label subjectLabelOne;
-
-  @FXML
-  private HBox hboxThree;
-
-  @FXML
-  private Label subjectLabelTwo;
-
-  @FXML
-  private HBox hboxFour;
-
-  @FXML
-  private Label subjectLabelThree;
-
-  @FXML
-  private HBox hboxFive;
+  private AnchorPane recentAnchorPane;
 
   @FXML
   private Label usernameLabel;
@@ -192,51 +119,15 @@ public class MainWindowController extends BaseController implements Initializabl
     viewFactory.showLoginWindow(stage);
   }
 
-  @FXML
-  void hBoxMouserClickedAction(MouseEvent event) {
-    int widthOfImages = 225;
-
-    if (event.getTarget() instanceof ImageView) {
-      // TODO Not sure if this way leads to a solution but looks hopeful
-      ImageView imageView = (ImageView) event.getTarget();
-      imageView.getImage().getUrl(); // This just returns null
-    }
-    // TODO fix for widths of variable size
-    int element = (int) event.getX()/widthOfImages;
-    try {
-      discoverAnchorPane.getChildren().clear();
-      viewFactory.embedSubjectWindow(discoverAnchorPane, account, subjectManager, element);
-    } catch (IOException e) {
-      e.printStackTrace();
-    }
-    primaryTabPane.getSelectionModel().select(1);
-  }
-
-
 
   @Override
   public void initialize(URL url, ResourceBundle resourceBundle) {
     updateAccountViews();
 
-    //Connecting Scroll Bar with Scroll Pane
-    mainRecentScrollBar.setOrientation(Orientation.VERTICAL);
-    mainRecentScrollBar.minProperty().bind(mainRecentScrollPane.vminProperty());
-    mainRecentScrollBar.maxProperty().bind(mainRecentScrollPane.vmaxProperty());
-    mainRecentScrollBar.visibleAmountProperty().bind(mainRecentScrollPane.heightProperty().divide(mainRecentScrollContent.heightProperty()));
-    mainRecentScrollPane.vvalueProperty().bindBidirectional(mainRecentScrollBar.valueProperty());
-
-    /* TODO Set Up Screen
-     * Request from server the top set of subjects.
-     * with each one get the server to send the thumbnail too.
-     * Fill out the display with the subjects and the thumbnails
-     *
-     * */
-    downloadSubjects();
-
-
     try {
       viewFactory.embedProfileWindow(popUpHolder, account);
       viewFactory.embedDiscoverWindow(discoverAnchorPane, account, subjectManager);
+      viewFactory.embedRecentWindow(recentAnchorPane, account, subjectManager, this);
     } catch (IOException e) {
       e.printStackTrace();
     }
@@ -269,37 +160,11 @@ public class MainWindowController extends BaseController implements Initializabl
     }
   }
 
-  private void downloadSubjects() {
-    //TODO Lots of error handling.
-    SubjectRequestService subjectRequestService =
-        new SubjectRequestService(getMainConnection(), subjectManager);
+  public TabPane getPrimaryTabPane() {
+    return primaryTabPane;
+  }
 
-    int subjectsBeforeRequest = subjectManager.getNumberOfSubjects();
-
-    if (!subjectRequestService.isRunning()) {
-      subjectRequestService.reset();
-      subjectRequestService.start();
-    }
-    subjectRequestService.setOnSucceeded(srsEvent -> {
-      SubjectRequestResult srsResult = subjectRequestService.getValue();
-
-      if (srsResult == SubjectRequestResult.SUCCESS || srsResult == SubjectRequestResult.FAILED_BY_NO_MORE_SUBJECTS) {
-        FileInputStream input = null;
-        for (int i = subjectsBeforeRequest; i < subjectManager.getNumberOfSubjects(); i++) {
-          try {
-            input = new FileInputStream(subjectManager.getSubject(i).getThumbnailPath());
-            Image image = new Image(input);
-            ImageView imageView = new ImageView(image);
-            imageView.setFitHeight(130);
-            imageView.setFitWidth(225);
-            hboxOne.getChildren().add(imageView);
-          } catch (FileNotFoundException e) {
-            e.printStackTrace();
-          }
-        }
-      } else {
-        System.out.println("Here in mainController " + srsResult);
-      }
-    });
+  public AnchorPane getDiscoverAnchorPane() {
+    return discoverAnchorPane;
   }
 }
