+++ client/src/main/resources/application/view/fxml/homeWindows/ScheduledWindow.fxml
